third_party/blink/web_tests/fast/forms/input-stepup-stepdown-expected.txt

Check stepUp() and stepDown() bahevior for unsupported types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Unsupported type
PASS input.step = "3"; input.min = ""; input.max = ""; input.value = "2"; input.stepDown() threw exception InvalidStateError: Failed to execute 'stepDown' on 'HTMLInputElement': This form element is not steppable..
PASS input.stepDown(0) threw exception InvalidStateError: Failed to execute 'stepDown' on 'HTMLInputElement': This form element is not steppable..
PASS input.stepUp() threw exception InvalidStateError: Failed to execute 'stepUp' on 'HTMLInputElement': This form element is not steppable..
PASS input.stepUp(0) threw exception InvalidStateError: Failed to execute 'stepUp' on 'HTMLInputElement': This form element is not steppable..
PASS successfullyParsed is true

TEST COMPLETE
